+++ lib/lite/command/base.rb
@@ -29,6 +29,8 @@ module Lite
 
       def initialize(context = {})
         @context = Lite::Command::Context.build(context)
+
+        on_pending if respond_to?(:on_pending, true)
       end
 
     end