+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added `on_success` callback
 - Added passing metadata to faults
+- Added `on_pending`, `on_executing`, `on_complete`, and `on_interrupted` callbacks
 ### Changed
 - Check error descendency instead of type
 ### Removed